+++ apps/whispering/src/lib/query/isomorphic/actions.ts
@@ -3,9 +3,9 @@ import { Ok } from 'wellcrafted/result';
 import { WhisperingErr } from '$lib/result';
 import { DbServiceErr } from '$lib/services';
 import { settings } from '$lib/stores/settings.svelte';
-import * as transformClipboardWindow from '../../routes/transform-clipboard/transformClipboardWindow.tauri';
-import { rpc } from './';
-import { defineMutation } from './_client';
+import * as transformClipboardWindow from '$routes/transform-clipboard/transformClipboardWindow.tauri';
+import { rpc } from '..';
+import { defineMutation } from '../_client';
 import { db } from './db';
 import { delivery } from './delivery';
 import { notify } from './notify';
