+++ apps/whispering/src/lib/query/isomorphic/db.ts
@@ -7,7 +7,7 @@ import type {
 	TransformationRun,
 } from '$lib/services/isomorphic/db';
 import { settings } from '$lib/stores/settings.svelte';
-import { defineMutation, defineQuery, queryClient } from './_client';
+import { defineMutation, defineQuery, queryClient } from '../_client';
 
 /**
  * Consolidated query keys that mirror the database service structure
@@ -252,7 +252,8 @@ export const db = {
 				const transformationsArray = Array.isArray(transformations)
 					? transformations
 					: [transformations];
-				const { error } = await services.db.transformations.delete(transformationsArray);
+				const { error } =
+					await services.db.transformations.delete(transformationsArray);
 				if (error) return Err(error);
 
 				queryClient.setQueryData<Transformation[]>(
