+++ apps/whispering/src/lib/query/isomorphic/delivery.ts
@@ -3,8 +3,8 @@ import { WHISPERING_RECORDINGS_PATHNAME } from '$lib/constants/app';
 import type { WhisperingError } from '$lib/result';
 import type { TextServiceError } from '$lib/services/isomorphic/text';
 import { settings } from '$lib/stores/settings.svelte';
-import { defineMutation } from './_client';
-import { rpc } from './index';
+import { rpc } from '..';
+import { defineMutation } from '../_client';
 
 export const delivery = {
 	/**
