+++ apps/whispering/src/lib/query/isomorphic/transcription.ts
@@ -1,14 +1,10 @@
 import { Err, Ok, partitionResults, type Result } from 'wellcrafted/result';
 import { WhisperingErr, type WhisperingError } from '$lib/result';
-import {
-	db as dbService,
-	transcriptions,
-	desktopServices,
-} from '$lib/services';
 import type { Recording } from '$lib/services';
+import { desktopServices, services } from '$lib/services';
 import { settings } from '$lib/stores/settings.svelte';
-import { rpc } from './';
-import { defineMutation, queryClient } from './_client';
+import { rpc } from '..';
+import { defineMutation, queryClient } from '../_client';
 import { db } from './db';
 import { notify } from './notify';
 
@@ -31,7 +27,7 @@ export const transcription = {
 		): Promise<Result<string, WhisperingError>> => {
 			// Fetch audio blob by ID
 			const { data: audioBlob, error: getAudioBlobError } =
-				await dbService.recordings.getAudioBlob(recording.id);
+				await services.db.recordings.getAudioBlob(recording.id);
 
 			if (getAudioBlobError) {
 				return WhisperingErr({
@@ -106,7 +102,7 @@ export const transcription = {
 				recordings.map(async (recording) => {
 					// Fetch audio blob by ID
 					const { data: audioBlob, error: getAudioBlobError } =
-						await dbService.recordings.getAudioBlob(recording.id);
+						await services.db.recordings.getAudioBlob(recording.id);
 
 					if (getAudioBlobError) {
 						return WhisperingErr({
@@ -181,7 +177,7 @@ async function transcribeBlob(
 		await (async () => {
 			switch (selectedService) {
 				case 'OpenAI':
-					return await transcriptions.openai.transcribe(
+					return await services.transcriptions.openai.transcribe(
 						audioToTranscribe,
 						{
 							outputLanguage: settings.value['transcription.outputLanguage'],
@@ -193,7 +189,7 @@ async function transcribeBlob(
 						},
 					);
 				case 'Groq':
-					return await transcriptions.groq.transcribe(
+					return await services.transcriptions.groq.transcribe(
 						audioToTranscribe,
 						{
 							outputLanguage: settings.value['transcription.outputLanguage'],
@@ -205,7 +201,7 @@ async function transcribeBlob(
 						},
 					);
 				case 'speaches':
-					return await transcriptions.speaches.transcribe(
+					return await services.transcriptions.speaches.transcribe(
 						audioToTranscribe,
 						{
 							outputLanguage: settings.value['transcription.outputLanguage'],
@@ -216,7 +212,7 @@ async function transcribeBlob(
 						},
 					);
 				case 'ElevenLabs':
-					return await transcriptions.elevenlabs.transcribe(
+					return await services.transcriptions.elevenlabs.transcribe(
 						audioToTranscribe,
 						{
 							outputLanguage: settings.value['transcription.outputLanguage'],
@@ -227,7 +223,7 @@ async function transcribeBlob(
 						},
 					);
 				case 'Deepgram':
-					return await transcriptions.deepgram.transcribe(
+					return await services.transcriptions.deepgram.transcribe(
 						audioToTranscribe,
 						{
 							outputLanguage: settings.value['transcription.outputLanguage'],
@@ -238,7 +234,7 @@ async function transcribeBlob(
 						},
 					);
 				case 'Mistral':
-					return await transcriptions.mistral.transcribe(
+					return await services.transcriptions.mistral.transcribe(
 						audioToTranscribe,
 						{
 							outputLanguage: settings.value['transcription.outputLanguage'],
@@ -253,7 +249,7 @@ async function transcribeBlob(
 				// // Pure Rust audio conversion now handles most formats without FFmpeg
 				// // Only compressed formats (MP3, M4A) require FFmpeg, which will be
 				// // handled automatically as a fallback in the Rust conversion pipeline
-				// 	return await transcriptions.whispercpp.transcribe(
+				// 	return await services.transcriptions.whispercpp.transcribe(
 				// 		audioToTranscribe,
 				// 		{
 				// 			outputLanguage: settings.value['transcription.outputLanguage'],
@@ -266,7 +262,7 @@ async function transcribeBlob(
 					// Pure Rust audio conversion now handles most formats without FFmpeg
 					// Only compressed formats (MP3, M4A) require FFmpeg, which will be
 					// handled automatically as a fallback in the Rust conversion pipeline
-					return await transcriptions.parakeet.transcribe(
+					return await services.transcriptions.parakeet.transcribe(
 						audioToTranscribe,
 						{ modelPath: settings.value['transcription.parakeet.modelPath'] },
 					);
@@ -274,7 +270,7 @@ async function transcribeBlob(
 				case 'moonshine': {
 					// Moonshine uses ONNX Runtime with encoder-decoder architecture
 					// Variant is extracted from modelPath (e.g., "moonshine-tiny-en" → "tiny")
-					return await transcriptions.moonshine.transcribe(
+					return await services.transcriptions.moonshine.transcribe(
 						audioToTranscribe,
 						{
 							modelPath: settings.value['transcription.moonshine.modelPath'],
